Fix decimal/double type mismatches in storage/task mapping

+++ TaskResourceBlueprints/Dto/ProjectTask/ProjectTaskListItemDto.cs
@@ -5,7 +5,7 @@
         string? Code,
         string DisplayName,
         string? UnitCode,
-        double? Quantity,
+        decimal? Quantity,
         bool IsActive
     )
     {
