+++ TaskResourceBlueprints/Dto/ProjectTask/TaskDefinitionEditDto.cs
@@ -19,7 +19,7 @@ namespace TaskResourceBlueprints.Dto.ProjectTask
         public int? UnitGroupId { get; set; }     // früher TaskUnitGroupId
         public string? UnitCode { get; set; }     // früher Unit
 
-        public double? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
 
         public double ChangeFactor1 { get; set; } = 1;    // früher ChangeFactor1
         public double ChangeFactor2 { get; set; } = 1;  // früher ChangeFactor2
@@ -29,9 +29,9 @@ namespace TaskResourceBlueprints.Dto.ProjectTask
         public string? Note { get; set; }
 
         // Deutsch: drei Eingabefelder (Dicke/Breite/Länge) – bequemer als Index-Binding auf eine Liste
-        public double? Thickness { get; set; } // WorkloadThresholds[0]
-        public double? Width { get; set; }     // WorkloadThresholds[1]
-        public double? Length { get; set; }    // WorkloadThresholds[2]
+        public decimal? Thickness { get; set; } // WorkloadThresholds[0]
+        public decimal? Width { get; set; }     // WorkloadThresholds[1]
+        public decimal? Length { get; set; }    // WorkloadThresholds[2]
 
         public List<int> VisibleFolderIds { get; set; } = new();    // früher VisibleFolders
